don't kill AC when ThwargFilter is not registered

+++ ThwargLauncher/ThwargLauncher/GameLaunching/GameLauncher.cs
@@ -220,7 +220,10 @@ namespace ThwargLauncher
             {
                 if (launcherProc != null && !launcherProc.HasExited)
                 {
-                    launcherProc.Kill();
+                    if (DecalInjection.IsThwargFilterRegistered())
+                    {
+                        launcherProc.Kill();
+                    }
                 }
             }
             if (launchResponse != null && launchResponse.IsValid)
